requirements.txt
lask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.0
Werkzeug==2.3.0
SQLAlchemy==2.0.0